+++ sound-mind-studio/tests/test_main_window.h
@@ -146,4 +146,6 @@ private slots:
     void paintingOnTheCanvasAppendsAPaintOperationToTheProjectsLog();
     void undoAndRedoDelegateToThePaintController();
     void settingANewProjectResetsPaintModeToOff();
+    void paintingWithTheDefaultToolConfigurationActuallyPaintsSomethingVisible();
+    void toggleToolConfigurationPanelShowsAndHidesIt();
 };