+++ sound-mind-studio/tests/test_tool_configuration_panel.h
@@ -1,0 +1,18 @@
+#pragma once
+
+#include <QObject>
+
+class ToolConfigurationPanelTest : public QObject {
+    Q_OBJECT
+
+private slots:
+    void freshPanelIsAnOpaqueCircularBrush();
+    void freshPanelHasBothOverlayCheckboxesOff();
+    void changingTheTipShapeEmitsToolConfigurationChanged();
+    void changingFalloffEmitsToolConfigurationChanged();
+    void changingSizeEmitsToolConfigurationChanged();
+    void changingIntensitySetsBothGradientStopsIntensity();
+    void changingOpacitySetsBothGradientStopsOpacity();
+    void togglingShowBoundingBoxesEmitsItsOwnSignal();
+    void togglingShowPathGeometryEmitsItsOwnSignal();
+};